Support testing global vars with
proper method interception
